feat: added currency-in-use deletion block

+++ src/main/java/com/jadebloom/goblin_api/currency/service/impl/CurrencyServiceImpl.java
@@ -2,7 +2,6 @@ package com.jadebloom.goblin_api.currency.service.impl;
 
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -10,29 +9,30 @@ import org.springframework.stereotype.Service;
 import com.jadebloom.goblin_api.currency.dto.CreateCurrencyDto;
 import com.jadebloom.goblin_api.currency.dto.CurrencyDto;
 import com.jadebloom.goblin_api.currency.entity.CurrencyEntity;
+import com.jadebloom.goblin_api.currency.error.CurrencyInUseException;
 import com.jadebloom.goblin_api.currency.error.CurrencyNameUnavailableException;
 import com.jadebloom.goblin_api.currency.error.CurrencyNotFoundException;
 import com.jadebloom.goblin_api.currency.mapper.CurrencyMapper;
 import com.jadebloom.goblin_api.currency.repository.CurrencyRepository;
 import com.jadebloom.goblin_api.currency.service.CurrencyService;
-import com.jadebloom.goblin_api.expense.service.ExpenseService;
+import com.jadebloom.goblin_api.expense.repository.ExpenseRepository;
 
 @Service
 public class CurrencyServiceImpl implements CurrencyService {
 
     private final CurrencyRepository currencyRepository;
 
-    private final ExpenseService expenseService;
+    private final ExpenseRepository expenseRepository;
 
     private final CurrencyMapper mapper;
 
     public CurrencyServiceImpl(
             CurrencyRepository currencyRepository,
-            @Qualifier("expenseServiceImpl") ExpenseService expenseService,
+            ExpenseRepository expenseRepository,
             CurrencyMapper mapper) {
         this.currencyRepository = currencyRepository;
 
-        this.expenseService = expenseService;
+        this.expenseRepository = expenseRepository;
 
         this.mapper = mapper;
     }
@@ -101,7 +101,14 @@ public class CurrencyServiceImpl implements CurrencyService {
     }
 
     @Override
-    public void deleteById(Long currencyId) {
+    public void deleteById(Long currencyId) throws CurrencyInUseException {
+        if (expenseRepository.existsByCurrency_Id(currencyId)) {
+            String f = "Cannot delete the currency with the ID '%d': some amount of expenses depend use it";
+            String errorMessage = String.format(f, currencyId);
+
+            throw new CurrencyInUseException(errorMessage);
+        }
+
         currencyRepository.deleteById(currencyId);
     }
